lamda changed type Signed-off-by: Yuxuan Zhang <y.zhang.4@student.tue.nl>

+++ Advanced Simulation - Assignment 3/Simulation/Simulation/Model/Queue/ContinuousQueue.java
@@ -5,7 +5,7 @@ import Simulation.Model.Queue.Behavior.ContinuousGenerateBehavior;
 
 public class ContinuousQueue extends Queue{
 
-	public ContinuousQueue(Queue_Priority queueingPriority, int minGroupSize, int rate, String queueID, String queueObjectName) {
+	public ContinuousQueue(Queue_Priority queueingPriority, int minGroupSize, double rate, String queueID, String queueObjectName) {
 		super(queueingPriority,  queueID);
 		generateQueueBehavior = new ContinuousGenerateBehavior(this.queueID, queueObjectName, rate,  queueObjects);
 	}
